requirements.txt
Django~=4.1.13
celery~=5.4.0
djangorestframework~=3.15.2
pip~=24.2
wheel~=0.43.0
redis~=5.0.7
amqp~=5.2.0
vine~=5.1.0
kombu~=5.3.7
billiard~=4.2.0
setuptools~=72.0.0
click~=8.1.7
python-dateutil~=2.9.0.post0
asgiref~=3.8.1
sqlparse~=0.5.1
colorama~=0.4.6
wcwidth~=0.2.13
six~=1.16.0
prompt_toolkit~=3.0.47
python-crontab~=3.2.0
django-timezone-field~=7.0